moved form option interfaces into form.ts
